requirements.txt
langchain==1.0.3
langchain-community==0.4.1
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
chromadb==1.3.0
openai==2.6.1
python-dotenv==1.2.1



